% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accurate.R
\name{accurate}
\alias{accurate}
\title{Accurate rate}
\usage{
accurate(da, tar, sigma, mu1, mu2, esti.mode = 0, cl.mode = 0)
}
\arguments{
\item{da, }{test dataset of a specified target group}

\item{tar, }{true value for the targets}

\item{sigma, }{estimated covariance matrix for the given target pair by Clime}

\item{mu1, }{mean values for the first target group}

\item{mu2, }{mean values for the second target group}

\item{esti.mode, }{digits for covariance estimating method, 0 for Kaul's method, 1/2/3 for Jun Li's 1st/2nd/3rd method}

\item{cl.mode, }{digits for classifying rule, 0 for the one in Kaul's software, 1 for the one in Kaul's paper}
}
\value{
the accuracy for the specific target group under given pair
}
\description{
Calculate accuracy rate for test dataset of a specified target group
}
