\name{clotu-package}
\docType{package}
\title{
\packageTitle{clotu}
}
\description{
\packageDescription{clotu}
}

\details{
Based on Kaul's assumption that non-zero OTUs follow a Gaussian distribution, this package incorporates the original model and  its mutations with modifications in normalizing OTU counts, shrinking number of OTUs and 3 mutated models of calculating covariances.
}

\author{
\packageAuthor{clotu}
}

\references{
A. Kaul, O. Davidov and S. D. Peddada, "Structural zeroes in high-dimensional data with applications to microbiome studies", Biostatistics, vol. 18, no. 3, p. 422-433, 2017. 

Jun Li, "Classification of microbiome data with structural zeroes and small samples", master thesis at Link\"{o}ping University, 2021
}

\keyword{ package }
\seealso{

}

\examples{

}
