% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvotu.R
\name{cvotu}
\alias{cvotu}
\title{Cross validation of OTUs}
\usage{
cvotu(da, ..., lambda = seq(0.001, 0.3, by = 0.01), nfold = 6, esti.mode = 0)
}
\arguments{
\item{da, }{dataset}

\item{..., }{mean values by esti.mode=2, available only when esti.mode=2}

\item{lambda, }{the range of lambda for cross validation, default value of seq(0.001,0.3,by=0.01)}

\item{nfold, }{number of folds for cross validation, default value of 6}

\item{esti.mode, }{digits for covariance estimating method, 0 for Kaul's method, 1/2/3 for Jun Li's 1st/2nd/3rd method}
}
\value{
the validation error for each iteration and lambda
}
\description{
Implement cross valildation and record the validation error for each iteration and lambda
}
