% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{data_split}
\alias{data_split}
\title{Split dataset}
\usage{
data_split(data, labels, vec)
}
\arguments{
\item{data, }{matrix of data}

\item{labels, }{character vector of labels for splitted datasets}

\item{vec, }{proportion of the splitted datasets}
}
\value{
list of splitted datasets
}
\description{
Split dataset into training/validation/test sets
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names
data_split(al,c("train","valid","test"),c(0.6,0.2,0.2))
}
