% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorcl.R
\name{majorcl}
\alias{majorcl}
\title{Major rules of classification}
\usage{
majorcl(tr, shrink.mode, esti.mode, cl.mode = 0, thr = 0, nvar = 75)
}
\arguments{
\item{tr, }{trained object by trotu function}

\item{shrink.mode, }{digits for shrinking method, 0 for Kaul's method, 1 for Jun Li's}

\item{esti.mode, }{digits for covariance estimating method, 0 for Kaul's method, 1/2/3 for Jun Li's 1st/2nd/3rd method}

\item{cl.mode, }{digits for classifying rule, 0 for the one in Kaul's software, 1 for the one in Kaul's paper}

\item{thr, }{threshold for zombie OTU}

\item{nvar, }{number of OTUs after shrinking}
}
\value{
Table of winning targets and its winning times
}
\description{
Implement major rules to classify an arbitrary observation against all targets
}
