% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muotu.R
\name{muotu}
\alias{muotu}
\title{Mean of OTU}
\usage{
muotu(da, esti.mode = 0)
}
\arguments{
\item{da, }{dataset}

\item{esti.mode, }{digits for covariance estimating method, 0 for Kaul's method, 1/2/3 for Jun Li's 1st/2nd/3rd method}
}
\value{
mean values for each OTU
}
\description{
Calculate mean value of target groups
}
