% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{normalizer}
\alias{normalizer}
\title{Normalizing OTU data}
\usage{
normalizer(data_tar, thr, method = 0, ref, del.otu = TRUE, del.sam = TRUE)
}
\arguments{
\item{data_tar, }{dataset split by targets}

\item{thr, }{threshold for zombie OTU/sample}

\item{method, }{digits for normalization method, 0 for Kaul's method, 1 for Jun Li's}

\item{ref, }{index of reference OTUs}

\item{del.otu, }{default value of T; if T the zombie OTU will be deleted}

\item{del.sam, }{default value of T; if T the zombie sample will be deleted}
}
\value{
normalized OTU dataset
}
\description{
Normalize OTU data with reference OTUs
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234) 
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names
ana0<-zomotu(al,0) ## delete otus not present in any sample, no otu names
ta<-target_split(da,ana0$otu,"Target")
ta_norm<-normalizer(ta,0,0,ana0$ref,FALSE,TRUE)
}
