% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomus.R
\docType{data}
\name{pomus}
\alias{pomus}
\title{Polish Mussel Dataset}
\format{
A list consisting of 2 data frames: otu, 693 OTUs by 108 samples; meta, information about 108 samples
\describe{
  \item{1st digit}{C for Czarna Hancza, P for Pilica river, S for San river}
  \item{2nd digit}{index of individuals}
  \item{3rd digit}{alias for different tissues: W for gut, G for gonads, M for muscles, N for foot, P for mantle, S for gills}
  \item{4th digit}{f for female, m for male}
}
}
\usage{
pomus
}
\description{
OTU table for 108 mussel samples from 3 Polish rivers: Czarna Hancza, Pilica river and San river. There are 693 OTUs, in particular, the microbial communities from gut, gonad, muscle, foot, mantle and gill tissues of 18 individuals from the 3 Unio crassus populations living in different ecological conditions are investigated. 3 male and 3 female individuals for each population were sampled in 2018 from 3 rivers in Poland with significantly distinctive ecological characteristics.
}
\keyword{datasets}
