% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present.R
\name{present}
\alias{present}
\title{Presency of OTUs}
\usage{
present(data)
}
\arguments{
\item{data, }{matrix of OTU data}
}
\value{
, proportion of non-zero counts for each OTU
}
\description{
Analyze the presency of OTUs
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names 
present(al)
}
