% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotu.R
\name{scotu}
\alias{scotu}
\title{Scaling OTUs}
\usage{
scotu(da, mu, esti.mode = 0)
}
\arguments{
\item{da, }{dataset}

\item{mu, }{mean values of OTUs}

\item{esti.mode, }{digits for covariance estimating method, 0 for Kaul's method, 1/2/3 for Jun Li's 1st/2nd/3rd method}
}
\value{
scaled counts for all observations
}
\description{
Scale dataset by substructing mean values
}
