% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_var.R
\name{shrink_var}
\alias{shrink_var}
\title{Split datasets}
\usage{
shrink_var(da1, da2, thr, nvar, pair = c(1, 2), method = 0)
}
\arguments{
\item{da1, }{normalized dataset for target 1}

\item{da2, }{normalized dataset for target 2}

\item{thr, }{threshold for zombie OTU}

\item{nvar, }{number of OTUs after shrinking}

\item{pair, }{character vector of specified target pairs, should be of length 2}

\item{method, }{digits for shrinking method, 0 for Kaul's method, 1 for Jun Li's}
}
\value{
a list of shrinked OTU tables for each target
}
\description{
Split dataset into training/validation/test sets
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names
ana0<-zomotu(al,0) ## delete otus not present in any sample, no otu names
ta<-target_split(da,ana0$otu,"Target")
ta_norm<-normalizer(ta,0,1,ana0$ref,FALSE,TRUE)
shrink_var(ta_norm[[1]],ta_norm[[2]],0,75,method=1)
}
