% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simotu.diri40.R
\name{simotu.diri40}
\alias{simotu.diri40}
\title{Simulation data with Gaussian distribution}
\usage{
simotu.diri40(
  n,
  p = 693,
  t = 3,
  ref = c(1, 214, 490, 512, 513),
  seed = F,
  unif.min = 0,
  unif.max = 0.35
)
}
\arguments{
\item{n, }{sample size for each target}

\item{p, }{number of OTUs}

\item{t, }{number of targets}

\item{ref, }{index of reference OTUs}

\item{seed, }{index of the seed, if F (default) no seed is set}

\item{unif.min, }{min limit of Uniform dist for non-zero probability}

\item{unif.max, }{max limit of Uniform dist for non-zero probability}
}
\value{
a list of OTU table and meta information
}
\description{
Generate simulation data with Dirichlet distribution
}
\examples{
simotu.diri40(36,p=693,t=3,ref=c(1,214,490,512,513),seed=1234,unif.min=0.9,unif.max=0.95)
}
