% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staotu.R
\name{staotu}
\alias{staotu}
\title{Statistics of OTU}
\usage{
staotu(da)
}
\arguments{
\item{da, }{matrix of OTU}
}
\value{
a list of mean and variance of those non-zero OTU counts, and index of non-zeros OTUs
}
\description{
Basic statistic values of each OTU
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names
staotu(al)
}
