% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suotu.R
\name{suotu}
\alias{suotu}
\title{Summary of classification accuracy}
\usage{
suotu(ha)
}
\arguments{
\item{ha, }{an returned object from function "trotu"}
}
\value{
summary of accuracies for each pair
}
\description{
Summary of accuracies for each target pair
}
