% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_split.R
\name{target_split}
\alias{target_split}
\title{Split into target groups}
\usage{
target_split(data_ori, data_pro, target)
}
\arguments{
\item{data_ori, }{original dataset consisting dataframes of "otu" and "meta"}

\item{data_pro, }{extracted matrix of specified condition}

\item{target, }{target variable}
}
\value{
a list of dataframes for each target
}
\description{
Split extracted data into different target groups
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names
ana0<-zomotu(al,0) ## delete otus not present in any sample, no otu names
ta<-target_split(da,ana0$otu,"Target")
}
