% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varotu.R
\name{varotu}
\alias{varotu}
\title{Variance of OTU}
\usage{
varotu(da, ..., esti.mode = 0)
}
\arguments{
\item{da, }{dataset}

\item{..., }{mean values by esti.mode=2, available only when esti.mode=2}

\item{esti.mode, }{digits for covariance estimating method, 0 for Kaul's method, 1/2/3 for Jun Li's 1st/2nd/3rd method}
}
\value{
covariance matrix of sample OTUs
}
\description{
Calculate sample covariance matrix for a given target pair
}
