% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zomotu.R
\name{zomotu}
\alias{zomotu}
\title{Zombie OTUs}
\usage{
zomotu(data, thr = 0, del = TRUE)
}
\arguments{
\item{data, }{matrix of OTU data}

\item{thr, }{threshold for zombie OTU}

\item{del, }{when T, delete zombie OTUs}
}
\value{
a list of preprocessed dataset, reference OTUs
}
\description{
Find zombie OTUs, and preprocess the dataset through deleting zombie OTUs and zombie observations
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
al<-data_extract(da,Target \%in\% c("target1","target2","target3")) # no otu names 
zomotu(al,thr=0)
}
