% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{data_extract}
\alias{data_extract}
\title{Extract data}
\usage{
data_extract(data, ...)
}
\arguments{
\item{data, }{a list consisting of 2 data frames: otu, each column is an observation and each row is a serie of OTU counts for all observations; meta, each row contains all information for that observation}

\item{..., }{filter condition for special subset of the given data}
}
\value{
a matrix of OTUs meeting the specified condition
}
\description{
Extract OTU data for specified condition
}
\examples{
da<-simotu.gaus(50,700,3,nref=5,full.mean=10000,unif.min=0,unif.max=0.4,seed=1234)  
ha<-data_extract(da,Target \%in\% c("target1","target2","target3")) ## return a matrix with 693 OTUs and 18 observations
}
